requirements.txt
# Display and GPIO - Compatible versions for Raspberry Pi
RPi.GPIO==0.7.1
gpiozero==1.6.2
Pillow==9.5.0
spidev==3.5
numpy==1.24.3

# Web Framework - Stable versions
Flask==2.2.5
Flask-CORS==4.0.0
Werkzeug==2.2.3
MarkupSafe==2.1.3

# System Utilities
psutil==5.9.5
netifaces==0.11.0

# AI Integration
openai==0.28.1
requests==2.31.0

# Network Management
python-wifi==0.6.1

# JSON and Config
jsonschema==4.17.3

# Additional utilities
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
